move RAG notebook to 254
move RAG notebook to 254
