+++ notebooks/254-llm-chatbot/config.py
@@ -9,7 +9,6 @@ You are a helpful, respectful and honest assistant. Always answer as helpfully a
 If a question does not make any sense or is not factually coherent, explain why instead of answering something not correct. If you don't know the answer to a question, please don't share false information.\
 """
 
-
 def red_pijama_partial_text_processor(partial_text, new_text):
     if new_text == "<":
         return partial_text
@@ -31,7 +30,7 @@ def chatglm_partial_text_processor(partial_text, new_text):
     return partial_text
 
 
-SUPPORTED_MODELS = {
+SUPPORTED_LLM_MODELS = {
     "red-pajama-3b-chat": {
         "model_id": "togethercomputer/RedPajama-INCITE-Chat-3B-v1",
         "remote": False,
@@ -49,6 +48,10 @@ SUPPORTED_MODELS = {
         "current_message_template": "{user} [/INST]{assistant}",
         "tokenizer_kwargs": {"add_special_tokens": False},
         "partial_text_processor": llama_partial_text_processor,
+        "prompt_template": """[INST]Human: <<SYS>> You are an assistant for question-answering tasks. Use the following pieces of retrieved context to answer the question. If you don't know the answer, just say that you don't know. Use three sentences maximum and keep the answer concise.<</SYS>> 
+        Question: {question} 
+        Context: {context} 
+        Answer: [/INST]""",
     },
     "mpt-7b-chat": {
         "model_id": "mosaicml/mpt-7b-chat",
@@ -65,7 +68,7 @@ SUPPORTED_MODELS = {
         "history_template": "<|im_start|>user\n{user}<im_end><|im_start|>assistant\n{assistant}<|im_end|>",
         "current_message_template": '"<|im_start|>user\n{user}<im_end><|im_start|>assistant\n{assistant}',
         "stop_tokens": ["<|im_end|>", "<|endoftext|>"],
-        "revision": "2abd8e5777bb4ce9c8ab4be7dbbd0fe4526db78d"
+        "revision": "2abd8e5777bb4ce9c8ab4be7dbbd0fe4526db78d",
     },
     "chatglm2-6b": {
         "model_id": "THUDM/chatglm2-6b",
@@ -75,6 +78,12 @@ SUPPORTED_MODELS = {
         "partial_text_processor": chatglm_partial_text_processor,
         "current_message_template": "[Round{num}]\n\n问：{user}\n\n答：{assistant}",
         "stop_tokens": ["</s>"],
+        "prompt_template": """基于以下已知信息，请简洁并专业地回答用户的问题。
+        如果无法从中得到答案，请说 "根据已知信息无法回答该问题" 或 "没有提供足够的相关信息"。不允许在答案中添加编造成分。另外，答案请使用中文。
+        已知内容:
+        {context}
+        问题:
+        {question}"""
     },
     "mistal-7b": {
         "model_id": "mistralai/Mistral-7B-v0.1",
@@ -84,7 +93,10 @@ SUPPORTED_MODELS = {
         "current_message_template": "{user} [/INST]{assistant}",
         "tokenizer_kwargs": {"add_special_tokens": False},
         "partial_text_processor": llama_partial_text_processor,
-        
+        "prompt_template": """<s> [INST] You are an assistant for question-answering tasks. Use the following pieces of retrieved context to answer the question. If you don't know the answer, just say that you don't know. Use three sentences maximum and keep the answer concise. [/INST] </s> 
+        [INST] Question: {question} 
+        Context: {context} 
+        Answer: [/INST]""",
     },
     "zephyr-7b-beta": {
         "model_id": "HuggingFaceH4/zephyr-7b-beta",
@@ -93,7 +105,6 @@ SUPPORTED_MODELS = {
         "history_template": "<|user|>\n{user}</s> \n<|assistant|>\n{assistant}</s> \n",
         "current_message_template": "<|user|>\n{user}</s> \n<|assistant|>\n{assistant}",
     },
-
     "neural-chat-7b-v3-1": {
         "model_id": "Intel/neural-chat-7b-v3-1",
         "remote": False,
@@ -102,7 +113,10 @@ SUPPORTED_MODELS = {
         "current_message_template": "{user} [/INST]{assistant}",
         "tokenizer_kwargs": {"add_special_tokens": False},
         "partial_text_processor": llama_partial_text_processor,
-        
+        "prompt_template": """<s> [INST] You are an assistant for question-answering tasks. Use the following pieces of retrieved context to answer the question. If you don't know the answer, just say that you don't know. Use three sentences maximum and keep the answer concise. [/INST] </s> 
+        [INST] Question: {question} 
+        Context: {context} 
+        Answer: [/INST]""",
     },
     "notus-7b-v1": {
         "model_id": "argilla/notus-7b-v1",
@@ -111,4 +125,8 @@ SUPPORTED_MODELS = {
         "history_template": "<|user|>\n{user}</s> \n<|assistant|>\n{assistant}</s> \n",
         "current_message_template": "<|user|>\n{user}</s> \n<|assistant|>\n{assistant}",
     },
+}
+
+SUPPORTED_EMBEDDING_MODELS = {
+    "all-mpnet-base-v2": {"model_id": "sentence-transformers/all-mpnet-base-v2"}
 }
